Refactoring the code and adding documentation
