+++ example/lib/src/app.dart
@@ -1,8 +1,7 @@
-import 'package:blendverse_ui_example/src/resources/values/app_theme.dart';
+import 'package:blendverse_ui/values.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-
 import 'settings/settings_controller.dart';
 import 'settings/settings_view.dart';
 import 'story/index.dart';
@@ -56,6 +55,7 @@ class MyApp extends StatelessWidget {
           // Define a light and dark color theme. Then, read the user's
           // preferred ThemeMode (light, dark, or system default) from the
           // SettingsController to display the correct theme.
+          //
           theme: AppTheme().lightTheme,
           darkTheme: AppTheme().darkTheme,
           themeMode: settingsController.themeMode,
