+++ lib/values.dart
@@ -1,0 +1,3 @@
+// Exports the values for AppColors and AppThemes
+export 'src/res/values/app_colors.dart';
+export 'src/res/values/app_theme.dart';
